requirements.txt
sherlock-project
python-dotenv
openai>=1.0.0
